+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [4.2.0](https://github.com/rizen/pixi-glyphs/compare/v4.1.8...v4.2.0) (2025-10-14)
+
 ### [4.1.8](https://github.com/rizen/pixi-glyphs/compare/v4.1.7...v4.1.8) (2025-10-13)
 
 - Fixed: PixiJS Deprecation Warning: Container.name property has been removed, use Container.label instead